Updates to calc
Now handles a variety of situations better.  Allows numbers that start with a decimal like .3, as well as numbers in scientific notation like 3e3, 3.e3, 3.3e3, .3e3, 3.3e+3, 3.3e-3, and all of those as negatives.  Handles a lot of edge cases better and returns errors for invalid inputs.  Also added unit tests to verify correctness in a wide variety of situations.  All unit tests are passed by this commit.